create join between stages and survey templates
