+++ task/seed.js
@@ -15,7 +15,7 @@ Promise.all([
     database.sequelize.model('question_option').bulkCreate(seedData.questionOption),
     database.sequelize.model('question_template').bulkCreate(seedData.questionTemplate),
     database.sequelize.model('survey_template').bulkCreate(seedData.surveyTemplate),
-    database.sequelize.model('join_trials_and_surveys').bulkCreate(seedData.joinTrialsAndSurveys),
+    // database.sequelize.model('join_trials_and_surveys').bulkCreate(seedData.joinTrialsAndSurveys),
     database.sequelize.model('join_surveys_and_questions').bulkCreate(seedData.joinSurveysAndQuestions),
     database.sequelize.model('stage').bulkCreate(seedData.stage)
 ])
